src: todo: add "add" feature
Currently, the only way to manage the tasks file is editing it directly.
In the next commit series, this will be changed. This first commit is to
allow to add a new task with a single command.

This command requires a new option "-a" and then two parameters "tag"
and "task". Tag is the category in which the task will be put on, while
"task" is a descriptive text that represents the task.

It will read the todo file and copy all its info to a new file, if the
corresponding tag is found, it will add the new task in the next line.
If the tag is never found, creates a new tag and puts the tag there.

The next reasonable feature is to mark tasks as done (remove it).