src: init: add check_file function
Make a new function to check if todo_file is currently according to the
defined pattern. This is important to make sure that the script doesn't
fail if there was some misconfiguration by the user.

This can be easily expanded in the future, but currently it does only
one kind of test in the file: if there is any tag written without any
tasks. If it finds any of these, it will written "Empty" in the next
line.

Signed-off-by: Lucas Cardoso dos Santos <lucascardoso910@gmail.com>