src: todo: add delete task function
Going forward in the currently main objective that is to give the user a
full control over the tasks only with the terminal interface, this adds
a function that allows the exclusion of a task given a task and a tag.

This also uses the recently created function `check_file` to make sure
that if the user removed the last task under a tag, it will be added a
"Empty" keyword.

Signed-off-by: Lucas Cardoso dos Santos <lucascardoso910@gmail.com>