src: todo: add list a specific tag feature
Add a new feature to the read_file function, allowing the user to pass a
parameter to select what tag to be shown, instead of showing all of them
all the time.

The function may be a little overcomplicated right now, but it is
working properly. If no tag is found, it doesn't print anything. But may
be good to work in a reestructuration in the future, when all the basic
features are already implemented.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>