src: converting script files to Python
Bash was kind of complicated to be dealing with long files. So, for this
reason, I decided to change the language to the operations files to
Python. I still use Bash in some files, the main files, assistant and
todo, because of the easy way it deals with options.

I considered also to use C or C++ for these files for its speed, but
once they would be needed to be compiled multiple times and I still
don't know how it is made in a more professional level, decided to opt
it out. Maybe some day I will reconsidered it.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>